requirements.txt
pandas
openai
langchain
langchain-openai
django
djangorestframework
